% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze fusionACS microdata}
\usage{
analyze(data, ..., by = NULL, fun = NULL, cores = get_cores())
}
\arguments{
\item{data}{Data frame. fusionACS microdata pseudo-sample returned by \link{assemble}.}

\item{...}{Formulas. Used to define the desired analyses. See Examples.}

\item{by}{Optional variable(s) that collectively define the set of population subgroups for which each analysis is computed. Can be a mix of geographic (e.g. census tract) and/or socio-demographic microdata variables (e.g. poverty status); the latter may be existing variables on disk or custom variables created on-the-fly via \code{fun()}. If \code{NULL}, analysis is done for the whole (national) sample.}

\item{fun}{Function. Optional function for modifying \code{data} prior to analysis. \code{fun(data)} is called internally and the requested analyses performed on the output.}

\item{cores}{Integer. Number of cores used for multithreading in \code{\link[collapse]{collapse-package}} functions. The default is one less than the total available cores.}
}
\description{
Analyze a fusionACS microdata pseudo-sample created by \link{assemble}. Efficiently computes means, medians, sums, proportions, and counts of specified variables, optionally across population subgroups.
}
\examples{
my.data <- assemble(
 variables = c(hincp, np, btung, totsqft_en, acequipm_pub, state_name, county10, tract10),
 year = 2019,
 respondent = "household",
 state_name == "Texas"
)

test <- analyze(
  data = my.data,
  mean_btung ~ mean(btung),
  ~median(totsqft_en),
  ~mean(acequipm_pub),
  by = list(np, c(state_name, county10, tract10))
)
}
