% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Assemble fusionACS microdata}
\usage{
assemble(
  variables,
  respondent,
  ...,
  year = "auto",
  directory = get_directory(),
  cores = get_cores()
)
}
\arguments{
\item{variables}{Vector of quoted or unquoted survey variables to return. In addition to \code{variables}, the output microdata also includes universal identifier variables (see Details).}

\item{respondent}{Character. Whether to return "household" or "person" microdata; i.e. the type of survey respondent. When \code{respondent = "household"}, any person-level \code{variables} return the response for the head of household (i.e. reference person). When \code{respondent = "person"}, any household-level \code{variables} are replicated for each person within a household.}

\item{...}{Optional expressions passed to \link[dplyr:mutate]{mutate} to create new columns, \link[dplyr:filter]{filter} to subset rows, or \link[dplyr:select]{select} to remove variables (usually after a mutate). See Examples.}

\item{year}{Either 'auto' (the default) or an integer vector specifying the year(s) of ACS-PUMS microdata to return. \code{year = "auto"} automatically sets a plausible value based on the requested \code{variables}.}

\item{directory}{Character. Path to the local fusionACS data directory. This is typically created automatically by \link{get_microdata}.}

\item{cores}{Integer. Number of cores used for multithreading in \link[arrow:arrow-package]{arrow} operations when assembling microdata. The default is one less than the total available cores.}
}
\value{
A keyed \link[data.table:data.table]{data.table} containing the requested \code{variables}, as well as the following universal variables (always returned):
\describe{
\item{year}{Year of the ACS-PUMS microdata observation.}
\item{hid}{Household ID. Along with \code{year}, this uniquely identifies each ACS-PUMS respondent household.}
\item{pid}{Person ID (if \code{respondent = "person"}). Along with \code{year} and \code{hid}, this uniquely identifies each ACS-PUMS respondent person.}
\item{weight}{The ACS-PUMS central observation weight.}
}
}
\description{
Assemble a fusionACS microdata pseudo-sample, given a user's requested variable(s), year(s), and respondent type. It also allows modification of the microdata via arbitrary expressions passed to \link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, and \link[dplyr:select]{select} and computed efficiently via \href{https://arrow.apache.org/docs/r/reference/acero.html}{Arrow dplyr queries}.
}
\examples{
# Load household income (hincp), household size (np), and state from ACS,
#  plus natural gas consumption (btung) and square footage (totsqft_en) from RECS,
#  plus pseudo-assignment of county and tract from UrbanPop.
# Nationwide household data for ACS year 2019
test <- assemble(variables = c(hincp, np, btung, totsqft_en, state_name, county10, tract10),
                respondent = "household")

# Same as above but for years 2017-2019 and with optional expressions used to:
# 1. Restrict to households consuming natural gas in the state of Texas
# 2. Create a new variable (btung_per_ft2) measuring consumption per square foot
# 3. Remove btung and totsqft_en after creating btung_per_ft2
test <- assemble(variables = c(hincp, np, btung, totsqft_en, state_name, county10, tract10),
                respondent = "household",
                btung > 0,
                state_name == "Texas",
                btung_per_ft2 = btung / totsqft_en,
                -c(btung, totsqft_en))
}
