% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionACS-package.R
\docType{package}
\name{fusionACS-package}
\alias{fusionACS}
\alias{fusionACS-package}
\title{fusionACS: To do.}
\description{
To do.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ummel.github.io/fusionACS/}
}

}
\author{
\strong{Maintainer}: Kevin Ummel \email{kevin.ummel@yale.edu}

}
\keyword{internal}
