% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cores.R
\name{get_cores}
\alias{get_cores}
\alias{set_cores}
\title{Get and set the number of cores}
\usage{
get_cores()

set_cores(n)
}
\arguments{
\item{n}{Number of cores to use.}
}
\value{
Number of cores to use.
}
\description{
Control the number of cores used by fusionACS for parallel processing. The default (set at package load) is one less than the total number of cores detected by \link[parallelly::availableCores]{availableCores}, which is generally accurate for both local machines and high-performance compute (HPC) clusters.
}
