% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_directory.R
\name{get_directory}
\alias{get_directory}
\alias{set_directory}
\title{Get and set the default fusionACS microdata directory}
\usage{
get_directory()

set_directory(dir)
}
\arguments{
\item{dir}{Path to fusionACS data directory.}
}
\value{
File path.
}
\description{
The user does not generally need to use these functions, since \code{\link{get_microdata}} sets the default directory automatically.
}
