% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{get_microdata}
\alias{get_microdata}
\title{Download fusionACS microdata}
\usage{
get_microdata(version = "latest", overwrite = FALSE)
}
\arguments{
\item{version}{Version (release date) of the desired GitHub release. Defaults to the latest (i.e. most recent) microdata release. Passed to \code{\link[piggyback:pb_download]{pb_download}} internally.}
}
\value{
Message to console if successful.
}
\description{
By default, the latest (i.e. most recent) \href{https://github.com/ummel/fusionACS/releases}{microdata release} is downloaded to the \link[rappdirs:user_data_dir]{user data directory}. The default directory path is updated automatically and accessible via \code{\link{get_directory}}.
}
\examples{
\dontrun{
get_microdata()
}
}
