% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Load fusionACS microdata dictionary}
\usage{
dictionary(directory = get_directory(), verbose = TRUE)
}
\arguments{
\item{directory}{Path to the local fusionACS data directory.}
}
\value{
A data frame with the following columns:
\describe{
\item{variable}{Variable name (column in microdata).}
\item{description}{Variable description or definition.}
\item{survey}{Acronym of the survey where the variable comes from.}
\item{vintage}{Vintage of the donor survey where the variable comes from (NA for ACS variables; see 'years' instead).}
\item{respondent}{Does the variable apply to "household" or "person" microdata observations?}
\item{type}{The data type of the variable (e.g. integer).}
\item{n_values}{Number of unique values the variable exhibits.}
\item{values}{Summary of the values found in the microdata. For numeric variables, returns the output of \code{\link[base]{summary}}. For categorical variables, returns a list of the unique values/levels.}
\item{years}{Year(s) of ACS-PUMS microdata for which the variable is available.}
\item{custom}{Logical indicating a 'custom' ACS variable created by the fusionACS team. See: https://github.com/ummel/fusionData/tree/master/survey-processed/ACS/custom}
}
}
\description{
Returns a data frame containing metadata for the microdata variables in the local fusionACS database.
}
\examples{
dict <- dictionary()
View(dict)
}
