% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionACS-package.R
\docType{package}
\name{fusionACS-package}
\alias{fusionACS}
\alias{fusionACS-package}
\title{fusionACS: Integrated, extensive, and high-resolution microdata of U.S. households and individuals}
\description{
This package provides access to data and tools to facilitate otherwise impossible U.S. social survey analyses. A documented pseudo-sample of the underlying fusionACS database is made available, along with functions to allow efficient microdata assembly and analysis. The fusionACS database is the result of statistical fusion of social survey microdata to simulate an extensive and high-resolution survey of U.S. households and individuals (Ummel et al. 2024, Tuccillo et al. 2023).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ummel.github.io/fusionACS}
  \item \url{https://github.com/ummel/fusionACS}
}

}
\author{
\strong{Maintainer}: Kevin Ummel \email{kevin.ummel@yale.edu}

Other contributors:
\itemize{
  \item Karthik Akkiraju \email{karthik.akkiraju@yale.edu} [contributor]
}

}
\keyword{internal}
